- some bugs Fixed

+++ src/main/java/Model/Zahra/Marshmallon.java
@@ -7,17 +7,18 @@ import Model.User;
 
 public class Marshmallon extends Monster {
     private Card card;
+
     public Marshmallon() {
         super("Marshmallon", Type.FAIRY);
     }
 
-    public void specialAction (boolean isNormalBattle, User rivalUser) {
-        long lifePoint = rivalUser.getLifePoint();
-       if (!isNormalBattle) {
-           Card marshmallon = (Marshmallon) card;
-           if (!marshmallon.getOccupied()) {
-               rivalUser.setLifePoint(lifePoint);
-           }
-       }
+    public void specialAction(boolean isNormalBattle, User rivalUser) {
+        int lifePoint = rivalUser.getLifePoint();
+        if (!isNormalBattle) {
+            Card marshmallon = (Marshmallon) card;
+            if (!marshmallon.getOccupied()) {
+                rivalUser.setLifePoint(lifePoint);
+            }
+        }
     }
 }
